cron/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
dotenv==0.9.9
requests==2.31.0
